Outsource selection component
